requirements.txt
# Requirements para el backend de IA de monitoreo sísmico
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
pandas>=1.5.0
scipy>=1.9.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.10.0
h5py>=3.7.0
requests>=2.28.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.6
flask>=2.3.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.10.0
cartopy>=0.21.0
shapely>=2.0.0
geopandas>=0.12.0
rasterio>=1.3.0
xarray>=2023.0.0
netcdf4>=1.6.0
tqdm>=4.64.0
python-dotenv>=1.0.0
schedule>=1.1.0
apscheduler>=3.10.0
pytest>=7.2.0
pytest-asyncio>=0.21.0
black>=22.0.0
isort>=5.10.0
flake8>=5.0.0
mypy>=1.0.0
jupyter>=1.0.0
ipykernel>=6.15.0
notebook>=6.4.0
jupyterlab>=3.4.0
docker>=6.0.0
kubernetes>=25.0.0
prometheus-client>=0.16.0
opentelemetry-distro>=0.40b0
opentelemetry-instrumentation-fastapi>=0.40b0